fix: setTVStandard signal to player

+++ optionstab.cpp
@@ -75,11 +75,11 @@ void OptionsTab::on_radioButtonPal_toggled(bool checked) {
     if (checked) {
         // Set to PAL
         pTrack->setTvMode(TiaSound::TvStandard::PAL);
-        emit setTVStandard(TiaSound::TvStandard::PAL);
+        emit setTVStandard(static_cast<int>(TiaSound::TvStandard::PAL));
     } else {
         // Set to NTSC
         pTrack->setTvMode(TiaSound::TvStandard::NTSC);
-        emit setTVStandard(TiaSound::TvStandard::NTSC);
+        emit setTVStandard(static_cast<int>(TiaSound::TvStandard::NTSC));
     }
 }
